# Improve performance of HTML parser on JVM